pyproject.toml
[project]
name = "fashion-concierge"
version = "0.1.0"
description = "Fashion Concierge AI agent using Google ADK and Gemini"
authors = [{name = "Fashion Concierge"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "google-generativeai[agent]>=0.5.0",
  "pydantic>=2.6.0",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]
